refactor(tasks): organizing and consolidating investor ubq distributor tasks
