+++ .gitignore
@@ -46,3 +46,6 @@ typings
 web/dapp
 deployments/*
 !deployments/mainnet
+contracts/dollar/tasks/library/distributor/fixtures/investors.json
+.gitignore
+contracts/dollar/tasks/library/distributor/fixtures/operators.json
