+++ contracts/dollar/tasks/library/distributor/bulk-transfer-to.ts
@@ -2,7 +2,7 @@ import "@nomiclabs/hardhat-waffle";
 import { Wallet } from "ethers";
 import "hardhat-deploy";
 import { ActionType } from "hardhat/types/runtime";
-import { impersonate } from "./investor-emissions/impersonate";
+import { impersonate } from "./bulk-transfer-to-library/impersonate";
 
 let distributor = "0x445115D7c301E6cC3B5A21cE86ffCd8Df6EaAad9";
